+++ include/ctre/phoenix/MotorControl/DeviceCatalog.h
@@ -13,7 +13,7 @@ public:
 		_mcs.push_back(motorController);
 	}
 
-	int MotorControllerCount() {
+	size_t MotorControllerCount() {
 		return _mcs.size();
 	}
 
@@ -33,6 +33,6 @@ private:
 };
 
 }
-} // namespace ctre_phoenix
+} // namespace phoenix
 }
